fix a region- and block- issue because of new negative numbers
When performing a minus operation, original positive numbers could be turned
into negative numbers, resulting in the introduction of new - signs. These
signs will change the beginning and the end positions of a region or a
block. This fix tracks possible increase of buffer size and use it to shift
the positions to the right.